Refactor get_text() method to Resume model and delete Docuemnt domain directory
