--- backend/tailor/domain/document/constants.py
@@ -1,5 +1,0 @@
-# File types
-DOCX = "DOCX"
-PDF = "PDF"
-TEX = "TEX"
-TXT = "TXT"
