--- backend/tailor/domain/document/tex.py
@@ -1,9 +1,0 @@
-from .base import Document
-
-
-class TexDocument(Document):
-    def get_text(self):
-        pass
-
-    def generate_copy(self):
-        pass
