--- backend/tailor/domain/document/txt.py
@@ -1,10 +1,0 @@
-from .base import Document
-
-
-class TxtDocument(Document):
-    def get_text(self):
-        with open(self.file.path, 'rb') as f:
-            return f.read()
-
-    def generate_copy(self):
-        pass
